Work/SSB/pyproject.toml
[tool.poetry]
name = "poetry-py3-8"
version = "0.1.0"
description = ""
authors = ["ChrisLee <ChrisLee@hyenatek.com>"]
readme = "README.md"
packages = [{include = "poetry_py3"}]

[tool.poetry.dependencies]
python = "3.8.10"
robotframework = "^6.1"
robotframework-pythonlibcore = "4.1.2"
pip = "23.1.2"
robotframework-requests = "0.9.5"
robotframework-seleniumlibrary = "6.1.0"
selenium = "4.9.1"
wheel = "0.38.4"
async-generator = "1.10"
attrs = "23.1.0"
certifi = "2023.5.7"
pyserial = "3.5"
setuptools = "67.8.0"
trio = "0.22.0"
apscheduler = "^3.10.1"
pyyaml = "^6.0.1"
pyinstaller = "^5.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
